no message
git-svn-id: svn+ssh://tictacserver.gel.usherbrooke.ca/Users/tictac/svnServer/developments/fullDiamant/diamant/trunk@1082 bc377d3f-2bd2-4e0d-9bbf-68e52cfa64df

+++ Dx/dInternal/dDataTxt/InstructorAttach.java
@@ -96,7 +96,7 @@ public class InstructorAttach extends DXObject {
    * INPUT: choice, an integer. choice =
    * OUTPUT: an integer. it return -1 if choice is invalid
    * */
-  public int getSelectedField(int choice){
+  public long getSelectedField(int choice){
     switch(choice){
       case 0: break;
     }
